3월 17일 회원제 수정 삭제

+++ MyFirstWeb/src/main/java/kr/co/ict/servlet/service/BoardDeleteService.java
@@ -2,6 +2,7 @@ package kr.co.ict.servlet.service;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.BoardDAO;
 
@@ -12,9 +13,15 @@ public class BoardDeleteService implements IBoardService{
 		
 		int board_num = Integer.parseInt(request.getParameter("board_num"));
 		
-		BoardDAO boarddao = BoardDAO.getInstance();
+		HttpSession session = request.getSession();
 		
-		boarddao.deleteBoard(board_num);
+		String writer = request.getParameter("writer");
+		String sId = (String)session.getAttribute("session_id");
 		
+		if(sId.equals(writer)) {
+			BoardDAO boarddao = BoardDAO.getInstance();
+			
+			boarddao.deleteBoard(board_num);
+		}
 	}
 }
